phân quyền

+++ controllers/admin/tour.controller.js
@@ -208,6 +208,13 @@ module.exports.trash = async (req, res) => {
     })
 }
 module.exports.createPost = async (req, res) => {
+    if (!permissions.includes("tour_create")) {
+        res.json({
+            code: "error",
+            message: "Bạn không có quyền sử dụng tính năng này!"
+        })
+        return;
+    }
     if (!req.body.position) {
         const count = await Tour.countDocuments();
         req.body.position = count + 1;
@@ -243,6 +250,13 @@ module.exports.changeMultiPatch = async (req, res) => {
         const { ids, status } = req.body;
         switch (status) {
             case "active":
+                if (!permissions.includes("tour_edit")) {
+                    res.json({
+                        code: "error",
+                        message: "Bạn không có quyền sử dụng tính năng này!"
+                    })
+                    return;
+                }
                 await Tour.updateMany({
                     _id: { $in: ids }
                 },
@@ -253,6 +267,13 @@ module.exports.changeMultiPatch = async (req, res) => {
                 req.flash('success', "Đổi trạng thái thành công!")
                 break;
             case "inactive":
+                if (!permissions.includes("tour_edit")) {
+                    res.json({
+                        code: "error",
+                        message: "Bạn không có quyền sử dụng tính năng này!"
+                    })
+                    return;
+                }
                 await Tour.updateMany({
                     _id: { $in: ids }
                 },
@@ -263,6 +284,13 @@ module.exports.changeMultiPatch = async (req, res) => {
                 req.flash('success', "Đổi trạng thái thành công!")
                 break;
             case "delete":
+                if (!permissions.includes("tour_delete")) {
+                    res.json({
+                        code: "error",
+                        message: "Bạn không có quyền sử dụng tính năng này!"
+                    })
+                    return;
+                }
                 await Tour.updateMany({
                     _id: { $in: ids }
                 },
@@ -311,6 +339,13 @@ module.exports.edit = async (req, res) => {
     }
 }
 module.exports.editPatch = async (req, res) => {
+    if (!permissions.includes("tour_edit")) {
+        res.json({
+            code: "error",
+            message: "Bạn không có quyền sử dụng tính năng này!"
+        })
+        return;
+    }
     try {
         const id = req.params.id;
         if (!req.body.position) {
@@ -354,6 +389,13 @@ module.exports.editPatch = async (req, res) => {
 
 }
 module.exports.deletePatch = async (req, res) => {
+    if (!permissions.includes("tour_delete")) {
+        res.json({
+            code: "error",
+            message: "Bạn không có quyền sử dụng tính năng này!"
+        })
+        return;
+    }
     try {
         const id = req.params.id;
         await Tour.updateOne({
@@ -375,6 +417,13 @@ module.exports.deletePatch = async (req, res) => {
     }
 }
 module.exports.undoPatch = async (req, res) => {
+    if (!permissions.includes("tour_trash")) {
+        res.json({
+            code: "error",
+            message: "Bạn không có quyền sử dụng tính năng này!"
+        })
+        return;
+    }
     try {
         const id = req.params.id;
         await Tour.updateOne({
@@ -395,6 +444,13 @@ module.exports.undoPatch = async (req, res) => {
 
 }
 module.exports.deleteDestroyPatch = async (req, res) => {
+    if (!permissions.includes("tour_trash")) {
+        res.json({
+            code: "error",
+            message: "Bạn không có quyền sử dụng tính năng này!"
+        })
+        return;
+    }
     try {
         const id = req.params.id;
         await Tour.deleteOne({
